moved gamemode and InputAdapter inside BaseUI
