+++ frontier/core/src/test/java/com/zhaw/frontier/screens/StartScreenTest.java
@@ -1,6 +1,5 @@
 package com.zhaw.frontier.screens;
 
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
 import com.badlogic.gdx.assets.AssetManager;
